+++ lib/paco/parse_error.rb
@@ -3,6 +3,8 @@ module Paco
   class Error < StandardError; end
 
   class ParseError < Error
+    attr_reader :ctx, :pos, :expected
+
     # @param [Paco::Context] ctx
     def initialize(ctx, expected)
       @ctx = ctx
@@ -11,16 +13,16 @@ module Paco
     end
 
     def callstack
-      @ctx.callstack
+      ctx.callstack
     end
 
     def message
-      index = @ctx.index(@pos)
+      index = ctx.index(pos)
       <<~MSG
         \nParsing error
         line #{index[:line]}, column #{index[:column]}:
-        unexpected #{@ctx.eof? ? "end of file" : @ctx.input[@pos].inspect}
-        expecting #{@expected}
+        unexpected #{ctx.eof? ? "end of file" : ctx.input[pos].inspect}
+        expecting #{expected}
       MSG
     end
   end